PixelBot
Software that scans all the pixels on the screen and triggers a click event when it finds a match" is generally used as an automation tool. This software scans the screen image and finds a match for a specific visual. When a match is found, the software triggers a click event and automatically performs a specific task. For example, it can be used as an automatic game player or a web scraping tool. This type of software typically works at the user's request and should not be used for illegal purposes.